Do not record inline stack metadata when not inlining
Summary:
If the frame is `-1`, we are in an unreachable code, so they will be never
queried. If the frame is `0`, we will record the parent frame as `-1`, which
matches `kInvalidFrameID`, which will be ignored by `inlineStackAt()`, which is
the only user of these data.

So, save the memory and do not record the inline stack metadata in these cases.

This diff also kills one place that uses `marker.bcOff()` in the prologue. We
do not ever want to read bytecode offsets in prologues.

Reviewed By: ricklavoie

Differential Revision: D24242953

fbshipit-source-id: 73a3a64fe7ea7b5577797f8a7bfe1f9ccdb1caed