Add no_pos_hash crate and implement for oxidized types
Summary:
We'd like the ability to hash decls in a way which isn't sensitive to changes in positions. It would be useful for decl-diffing in incremental typechecks, and for producing decl hashes for the file info table.

A simple way of doing this with the `std::hash::Hash` trait would be to write `impl Hash for Pos` as a no-op. This is a misuse of the `Hash` trait, however--[non-equal positions should not hash to the same value](https://doc.rust-lang.org/std/hash/trait.Hash.html#hash-and-eq). As a result, we wouldn't be able to do things like store positions in a HashSet (which seems like a reasonable thing to do).

Another approach might be replacing positions with `Pos::none()` before hashing them (as we do in OCaml). For by-ref types, this would involve an endo-reallocation, and we don't yet have an endo-visitor for our by-ref types.

Instead, this diff adds a new hashing trait `NoPosHash` and a procedural macro to derive implementations of it. The trait is expected to behave the same way as `Hash`, except its implementation for `Pos` is a no-op. This way, `Hash` still behaves correctly for positions, and we need not implement an endo-visitor or perform allocations when we need a position-insensitive hash.

Reviewed By: shiqicao

Differential Revision: D24228808

fbshipit-source-id: 34c8e14348d3779e43ca02fa56f5e3fc9983c359