Guard collections updates with setArrayData()
Summary:
There's a lot of code in the collections implementation that uses m_arr (the ArrayData backing a collection) directly. We want to enforce invariants on these arrays. We do so by adding a setter that checks the invariants and using it everywhere. As a consequence, we also have to fix const-correctness in dataAt(), which we do by splitting it into lvalAt() and rvalAt().

This change is pure code motion. Nothing interesting happening here.

Reviewed By: dneiter

Differential Revision: D24262090

fbshipit-source-id: d781d054a9968f48d888b2fb17acc9df8160800b