Clean up onStrobelightSignal()
Summary:
This removes TRACE() calls (which are not safe) from the signal handler, and prunes some unused logic.

kSignumAll is left as an option, in case future profiling tools (ie a wall-time profiler that periodically dumps every active request stack) require it, but it is unused currently and has a potential race condition. Comment it out with a note about what would need to be fixed before it can be re-enabled.

Reviewed By: ricklavoie

Differential Revision: D24264054

fbshipit-source-id: f63c7e0cef65fcc0233c3759ad3b60c2086fd925