Simplify array_intersect_key
Summary:
array_intersect_key has a "fast path" that involves keeping track of lots of iterator indices in a vector. I'm not convinced that this path is any faster than the default path; the default path uses iteration over one array and (potentially hash) lookups on the other array, just like the fast path.

What this path does do is that it's the only user of the "GetIntPos" / "GetStrPos" APIs. We have no need of these.

Reviewed By: ricklavoie

Differential Revision: D24218892

fbshipit-source-id: 7489eca5c87ff6b71a3f886811c6eaff7be390b4