Special case subtyping for `mixed <: [unresolved]`
Summary:
In the case of `mixed <: Tvar v`, add `mixed = Toption Tnonnull` as a lower bound of `v`, rather than separately adding `null` and `nonnull` as lower bounds. This means that later, when an upper bound might be added to `v`, transitive closure executes `mixed <: <upper bound>` rather than separately `null <: <upper bound>`and `nonnull <: <upper bound>`.

This results in clearer error messages, referring to a failure to supertype `mixed`, rather than a failure to supertype `null`/`nonnull`. For example in bad method invocations on mixed values which trigger this mechanism, the user is now clearly informed that they should not invoke a method on a `mixed` value. Such a clearer error message is shown in test `super_constraint_9.php`.

Reviewed By: CatherineGasnier

Differential Revision: D24217334

fbshipit-source-id: 69ae4c545ec7868df4111e13bdb268fb68a66779