Fix unsoundness with new static
Summary:
I've been trying to clean up our code around type constants, expression dependent types and the `this` type (since they are all related currently) and uncovered some unsoundness with how we currently type `new static()`. Here is some code that type checks today that shouldn't:

```hack
<<__ConsistentConstruct>>
abstract class C<T> {
  public static function getInstance(): C<int> {
    return new static();
  }

  abstract public function get(): T;
}

class Cint extends C<int> {
  public function get(): int {
    return 0;
  }
}

class Cstr extends C<string> {
  public function get(): string {
    return '';
  }
}

function breakit(): int {
  $cint = Cstr::getInstance();
  return $cint->get();
}
```

The problem stems from the usage of `new static()` in `C::getInstance`. The type we give in this case to the expression is roughly: `this(C<#1>)`. The `this(...)` is meant to tag that it refers to the late static bound type, but the real issue here is we instantiate `C` with a fresh type variable. This is unsound because a subclass of `C` may have instantiated the generic with a different type (like is the case for `Cstr`).

Instead of instantiating with a fresh type variable, we should  instantiate `C` with the generic `T`. Then we would get an error when returning `new static()` because `C<T> <: C<int>` is unsatisfiable. This is essentially what this diff does.

It isn't quite obvious it does this because of how complex `Typing.new_object` is, but `c_ty` is the type derived from calling `static_class_id`. For `CIstatic` this will return the proper.

Reviewed By: DavidSnider

Differential Revision: D24270952

fbshipit-source-id: 6d0416fbbbeeec2f7e2306412c43fb4ff48206e6