remove invalid hot decls
Summary:
If we do a saved-state init and then process dirty files, and those dirty files mean that some of the hot decls are no longer valid, then we need to remove them!

Prior to this diff, we weren't removing them. We were merely re-typechecking the fanout of the dirty files, which includes the files with the hot decls. But while typechecking the fanout, nothing stopped us from reading those invalid hot decls, and so we got Decl_heap_elems_bug.

Differential Revision: D24273713

fbshipit-source-id: f674763376673ae3c58af8288955441565681b4d