Iterate over small set instead of large set
Summary:
When a base class changes, we collect the files in which its subclasses are declared, then check the decl of each class declared in that file to see if the base class is one of its ancestors. If it is, we invalidate that decl. This check allows us to avoid unnecessarily invalidating decls for classes that happen to be in the same file as one that needs invalidation.

However, the way we check if the class's ancestors are in the set of base classes which changed seems inefficient. We iterate over the set of changed classes, and for each, test for membership in the set of ancestors for the candidate subclass. This seems okay for redecl phase 1, where the set of changed classes is expected to be very small, but might be very expensive in redecl phase 2, where the set of changed classes has expanded to include all subclasses.

This whole system is intended to be removed after we've shipped shallow decl, so I don't think we should invest too much effort here. But I think this `intersection_nonempty` check can easily be changed to perform better in phase 2--we just ought to iterate over the smaller collections of ancestors instead of the large collections of changed classes.

Reviewed By: DavidSnider

Differential Revision: D24260164

fbshipit-source-id: bd785c87b1f935741fa3023d2816065bd3bc1222