Fix bug in unit-cache when re-using Units with unchanged SHA1s
Summary:
When we detect a file has changed, we'll reparse the file to obtain a
new Unit. However if the file's SHA1 hasn't actually changed (for
example of it was just touched), we'll skip this and just reuse the
existing data. We'll do the same if we detect the BC SHA1 hasn't
changed after creating a new Unit. The rational is that its better to
keep an existing Unit, and any JITing we may have done with it.

However this logic has one bug. We reuse the existing entry entirely,
which means we don't update the stat information (the stat information
is by definition changed since we detected a file change). As a
result, we'll that entry will continue to be treated as "changed" from
that point onward. In fact, every request afterward (until the Unit
actually changes), we'll detect the Unit has changed, and fall back
into the compile path.

Fix this by creating a new CachedUnitWithFree entry for the Unit
instead of reusing the old one. This lets us reuse the Unit while
updating the stat information. However, this opens up a different
problem. It means different CachedUnitWithFree instances can "own" the
same Unit and neither knows when its safe to delete it. Add a
ref-count to UnitExtended which allows us to deal with this. The
ref-count indicates how many CachedUnitWithFree instances hold the
Unit. Only the one which drops the ref-count to zero actually deletes
it. This same mechanism will actually be needed for re-using Units by
hash anyways.

Since we store the ref-count in UnitExtended it shouldn't affect prod
at all.

Reviewed By: jano

Differential Revision: D24292000

fbshipit-source-id: a4fe413f7446928896757ef6895ee4ae95961fa6