Refine function (parameter) components in constructed supertype for new function type checking
Summary:
In the case of `call` in which the function is unresolved, refine the flags and components used to construct a function supertype, and clean up comments.

We would like the supertype to be as flexible and permissive as possible, so we choose flags that permit this (according to `subtype_funs`).

In particular:
- supertype is always non-reactive
- parameters are maybe mutable
- parameters do not have defaults
- parameters have no `Rx` annotations

The following appear to have no impact on subtyping:
- type parameters
- `where` constraints
- enforceability

Lastly, the following are incomplete and must be handled in future work:
- `inout` parameters
- parameters accepting disposables
- functions returning disposables

Reviewed By: CatherineGasnier

Differential Revision: D24081193

fbshipit-source-id: 5652dea2089efb933a2493539d83f092c5a001f7