Faster inc/dec ops on MonotypeVec
Summary:
We can use the "monotype" constraint to avoid a lot of branches in these operations. We split into three cases: for an uncounted type, we do nothing; for a guaranteed-counted type (e.g. Object), we do branchless incs, and for a maybe-counted type (e.g. Array) we do checked incs. Similarly, we optimize decs.

We also pull out a typedValueUnchecked() helper which is used in several places.

There's more to do with refcounting - primarily, we need to add the fast paths for constructing a MonotypeVec by taking ownership of stack values. But we leave it as a follow-up.

Differential Revision: D24298232

fbshipit-source-id: 679cedb55ef331437d65b7e6265b17566dc06049