Ensure key is known DataType in irgen-minstr
Summary:
Throughout irgen-minstr, it is assumed that we have a DataTypeSpecific for the key. This is used to determine whether to dispatch to TInt or TStr cases (or to perform class to string coercion). Surprisingly, we don't seem to enforce this. If we wind up spitting out some InterpOnes earlier in the tracelet, we might not have this type information and irgen-minstr will do the Wrong Thing.

This issue occurred later in the stack when we punted on Cls to string coercion cases, but it seems like it could occur more generally without these checks.

Reviewed By: ricklavoie

Differential Revision: D24309862

fbshipit-source-id: a22411dffc39820ce5f09eae25781df2d217cf57