Fix memory effects for NewLoggingArray
Summary:
When NewLoggingArray was emitted in non-profiling tracelets for testing, several assertion failures cropped up where the value returned by VMRegAnchor was incorrect. The root cause appears to be that when inlining is performed along with store elimination, if there is no dependency on components of the ActRec (such as the function pointer used by VMRegAnchor), they will not be materialized. This store elimination was occurring on the inlined ActRec's function pointer. This results in a correct fixup for the inlined callee, but an incorrect function pointer corresponding to the caller.

To recify this, we make the dependency on the current fp explicit in `memory-effects.cpp`.

Note that this appears to be a more general problem: anything that relies on the accuracy of VMRegAnchor should be specifying ActRec memory dependencies to prevent them from being elided. As we rely on this functionality for every logging operation of LoggingArray, we perform a sanity check on the VMRegAnchor results before using them to construct a SrcKey and potentially causing an assertion failure. Short of marking IR ops like DecRef (which can dispatch to LoggingArray release and trigger a VMRegAnchor) as dependent on ActRec information, I'm not sure how we can more generally fix this.

Differential Revision: D24271894

fbshipit-source-id: 92ebefd312ed4edb0dc6fd543d303fe8c0f0f96e