Custom dispatch for bespoke implementations of layout-sensitive bytecodes
Summary:
When translation dispatches to `handleBespokeInputs`, we now dispatch to a translator used only for layout-sensitive bytecodes used with bespoke arrays. This allows us to leave the current vanilla irgen untouched and isolate the bespoke implementations. This diff implements the bespoke side of translation dispatch and allows bespoke types to reach live tracelets.

We also use this dispatch path to JIT code for logging arrays in profiling tracelet diamonds. Because of this, we remove the guards on `AllowBespokesInLiveTypes` as it is no longer necessary.

Differential Revision: D24261426

fbshipit-source-id: 47e69086beb0d6e4951786adc98a9d8999416822