Fix prefetcher crash
Summary:
D24292000 (https://github.com/facebook/hhvm/commit/82d766e06158a8f250078e6bb600a7cfc34e7c3c) added an assert that we were never trying to update
the unit-cache with the same entry as before. However this can happen
in the Unit prefetcher for the non-canonical path. We already guard
against this in the main Unit loading path. Do the same in the
prefetcher.

Reviewed By: oulgen, mofarrell

Differential Revision: D24322098

fbshipit-source-id: 2789a4ca81545c5ae27db44503a248cf21030be0