Fix post-diamond assertions to avoid trashing typeSrcs
Summary:
The way we were previously emitting assertions after logging diamonds was not correct. Effectively, an AssertLoc/AssertStk has no typeSrc in the frame state (other than itself), and so any constraint created afterward that is traced back to this Assert will not cause a guard to be strengthened. Therefore, we emit the assertions on the logging side of the diamond instead of after it, so that typeSrcs can be traced back through the vanilla side. Without this fix, the subsequent diff (in which we make region formation and true IR generation identical) produces incorrect guard constraints.

I still anticipate removing these assertions entirely once bespoke translation is complete.

This diff is separate to ensure that the new assertion output is still compatible with vanilla region formation.

Differential Revision: D24307234

fbshipit-source-id: 9cb79feb8c58143711a654afe4282f01bb0daa77