Type expression tree as the result of inferring the desugared code
Summary:
Right now, in the TAST, we are typing the original expression within the expression tree and assigning that to the ExprTree node. Thus:

```
Code`4 + 4`
```
is typed as an int.

This is not what the type of this expression should be. Instead, we should type the result of this expression as the result of typing the desugared code. So we should think that

```
Code`4 + 4`
```
is an `ExprTree<..., ...>`

Reviewed By: Wilfred

Differential Revision: D24281227

fbshipit-source-id: 665a3ee2db5e54a7d63b3b535fa4f85c369bc8c7