Add a thunk and new ExprTree to desugaring
Summary:
We currently desugar to a lambda that users need to invoke with the visitor.

```
Visitor`2 + __splice__(3)`
```
Is currently desugared to:

```
(Visitor $v) ==> {
  // Extracted Splices
  return desugared_et;
}
```

We want to set up a two layered system of anonymous functions. So we want to desugar

```
Visitor`2 + __splice__(3)`
```

to

```
(() ==> {
  // Extracted Splices
  return new ExprTree(
    (Visitor $v) ==> { return desugared_et; }
  );
})()
```

Thus this diff adds the outer thunk and moves the extracted splice assignment there and also adds the invocation of the `new ExprTree(...)`.

Reviewed By: Wilfred

Differential Revision: D24196532

fbshipit-source-id: 40878bafd03290d2ba1b99d85067a7a1e16b1df0