Desugar splices into assignments
Summary:
Splices should be lifted to be executed before the ExprTree itself is created. This diff adds a Visitor that takes an expression, replaces all of the spliced expressions with temporary variables and adds assignments at the top. Thus, the visitor does the following desugaring:

```
Visitor`3 + __splice__(4)`
```
and converts to
```
(Visitor $v) ==> {
  $__0 = 4;
  return $v->add(
    $v->intLiteral(3),
    $v->splice($__0),
  );
}
```

Reviewed By: Wilfred

Differential Revision: D24166179

fbshipit-source-id: a473fdec6e9b3ebeedb206d22b5b281890e61167