Eliminate dependent_type.DTcls
Summary:
A couple years ago andrewjkennedy added an Exact/Nonexact flag to each class type. This makes the `dependent_type.DTcls` variant mostly redunant. Currently it is only constructed inside Typing_dependent_type and shouldn't leak outside that file since when we apply a DTcls it will set a class type to be exact.

So I made a small copy of dependent type inside Typing_dependent_type and then removed DTcls from dependent_type. This allows most of the logic to remain the same, but to kill off some should be dead code that was pattern matching against DTcls.

Reviewed By: andrewjkennedy, vassilmladenov

Differential Revision: D23984913

fbshipit-source-id: 1aa43a8c8d04487629108c37cf1840dd64d5dd4d