Allow Unit filepath's to be per-request
Summary:
We want to be able to reuse Units across files with the same hash. The
main blocker for this is that a Unit stores its filepath within it
(therefore two files with other identical contents cannot be
represented by the same Unit). In order to enable this, add the
ability for a Unit's filepath to be per-request.

When the Eval.ReuseUnitsByHash option is enabled, we store a Unit's
filepath in a normal RDS slot. When we load the Unit out of the
unit-cache, we bind the requested path into the slot. The usual
filepath() accessor will use the per-request filepath if available and
use the original filepath if not. A new accessor to always access the
original filepath is made available (this is used when the path is
being used as a key and should not vary).

This required a small change in the irgen for the File and Dir hhbc
opcodes. Previously we could always generate constants for them. We
continue to do so if the Unit doesn't have pre-request filepaths
enabled. Otherwise use new irgen instructions to load the path out of
RDS.

We explicitly stored the dirpath in the Unit, even though it was
always filepath with the file stripped off. Don't store this and strip
the file off the filepath on demand.

Slight changes are needed to the unit-cache logic. It needs to bind
the path when returning the Unit. However if the load is due to
prefetching, we do not want to bind the path (because the Unit isn't
going to be used). Also, when loading an existing Unit out of the
cache, we need to check if the path isn't already bound (or if bound
is the same as the requested path). If not, we need to treat it as a
miss and compile a new Unit (this can't happen currently).

Eval.ReuseUnitsByHash is only used in non-RepoAuthoritative mode, so
this shouldn't make any difference for production. The RDS slot itself
is stored in UnitExtended so does not consume memory when otherwise
not needed.

Reviewed By: paulbiss

Differential Revision: D24272798

fbshipit-source-id: 0c713f94d9dbe3366cf0dad2e40d6e12317cea8e