show e.g. "busy [--type-at-pos]" in busy status message
Summary:
The goal of this diff is so that when the user types `hh --type-at-pos` (for instance), and hh_server is therefore busy doing type-at-pos, then anyone else who invokes hh will see a status message
```
hh_server busy [--type-at-pos]
```
The idea is to show, in a human-relevant way, what command is currently being executed.

My philosophy is that the best place to come up with that human-relevant description, is the code that issues/parses the command in the first place.

Therefore this diff changes clientArgs so it sets a new field `args.desc` which is the option (--type-at-pos, ...) which caused the current command mode to be set.

Reviewed By: dabek

Differential Revision: D24240185

fbshipit-source-id: 34b59f76da61c1b43ad514ce97d0322ef7a0c510