tests shouldn't pollute /tmp/hh_server
Summary:
Our tests were sticking hundreds of Mb into /tmp/hh_server. This diff fixes it.

(Actually, the tests are written so they do leave junk in /tmp if they FAIL. That's so it's easier to investigate from the still-existing files. They only clean themselves up when they succeed. This is by design).

Reviewed By: vsiles

Differential Revision: D24285394

fbshipit-source-id: de60a5a06a24b91eb5164c2c0b4904639022c591