Type closure definition properly
Summary:
Closures (`Lfun` or `Efun`) also need to type their coeffects;
the annotations could be either explicit (in which we should look up the AST),
or implicit (i.e., inheriting coeffects from the enclosing scope).

In the former case, read the type off the newly inserted env binding.
In the latter case, look up the type of capabilities in the environment.

 ---

Unrelatedly, do a minor cleanup so that we don't get spurious OCaml errors when
using `hint_of_type_hint` and `type_hint_option_map` by making it less general:
check that the second element of the pair is in fact a hint. For the same reason,
add a missing helper function that is a more specialized version of `fst`.

Reviewed By: vassilmladenov

Differential Revision: D24211818

fbshipit-source-id: 5da4eed959fcac3421c85016629592a0d56fbdcd