More refinement exports on TAST
Summary:
Export refined environment in `Assert_env` blocks at while loop exit, if branches, and foreach statements.

## If statement

For an if statement,
```
if (...) {
  ...
} else {
  ...
}
```
we emit a refined environment in each branch:
```
if (...) {
  Assert_env refinement
  ...
} else {
  Assert_env refinement
  ...
}
```

## While statement

For a while loop we already emit refinement and join environments in the body,
```
while (...) {
  Assert_env refinement
  Assert_env join
  ...
}
```
we now also emit a refined environment for when the exit condition holds.
```
while (...) {
  Assert_env refinement
  Assert_env join
  ...
}
Assert_env refinement
```

This means the TAST of a while statement is a block statement that has the while statement and the refined environment in it.

## Foreach statement

`foreach` loops seem not to do any refinement, so we only export the join environment a similar manner to while loops.

```
foreach (...) {
  ...
}
```
becomes
```
foreach (...) {
  Assert_env join
  ...
}
```

## Other changes

Annotation API is changed slightly to be more generic.

There are changes to the Hack IFC test cases, this is because it was using fake members (which weren't paramount to the tests) and Hack IFC does not currently support them.

Reviewed By: mpu

Differential Revision: D24016952

fbshipit-source-id: e8db0777d61bf9bc279491d710b82534c9612264