Support for foreach statements
Summary:
Suport for foreach statements. This is very similar to while loops with the difference that we bind array key and values into distinct handles.

The implicit condition on the loop body being run is dependent on the size of the array, as such that is what flows into the pc.

Reviewed By: mpu

Differential Revision: D24189905

fbshipit-source-id: 1c15ede8bc686fbce558fbd67470f04d50517115