Implement SHA1 hash based Unit cache
Summary:
Right now we only cache Units according to their associated
paths. This works well if you're working on a typical source control
checkout, where pulling updates the files in place (or if you're
editing them yourselves). Once this happens, we detect that the file
has changed, we recompile it, and then update the cached entry. If the
file hasn't changed, we just use it as is. However, if one is managing
the files via a symlink, things don't work as well. For example: the
source tree is linked via a symlink to directory foo1. All the
canonicalized paths in the Unit cache have foo1 in their path (since
we resolve symlinks). The symlink is atomically updated to point to
directory foo2. Now, instead of detecting that the files have changed,
we simply treat them as brand new files (because their canonical paths
have foo2 in them). Even if the majority of files haven't changed,
we'll still recompile them (and also re-JIT them). Even worse, the old
Units will sit around in memory.

The key idea is that in this scenario, most of the "new" files will
have identical SHA1 hashes of the old. Therefore if we cache the Units
by hash as well, we can quickly reuse them instead of recompiling
everything. This is implemented by a standard cache behind the normal
path cache. In the symlink scenario we'll continue to treat it as a
new file, but we'll reuse the same Unit as the old one. Making the
paths in a Unit per-request unlocks this.

This is fairly straightforward. The only real difficulty is properly
removing entries from the hash cache when the last reference in the
path cache goes away.

Reviewed By: paulbiss

Differential Revision: D24321549

fbshipit-source-id: 28c51d44ba1bed06f349b0863e53dc6c91b775db