Desugar boolean `&&` and `||`
Summary:
Since we allow booleans to be desugared, we can allow boolean operators `&&` and `||`

`expr1 && expr2` is desugared into
```
$v->ampamp(expr1_desugared, expr2_desugared)
```

`expr1 || expr2` is desugared into
```
$v->barbar(expr1_desugared, expr2_desugared)
```

Naming can be changed. Currently just going based on the names in ast_defs as we've been attempting to name based on syntax rather than semantics.

Reviewed By: Wilfred

Differential Revision: D23946592

fbshipit-source-id: d6bfb6fdb0be94b8440cd71ca493cca8acd8c569