Initial monotype dict implementation
Summary:
We've got 200 lines of EmptyMonotypeDict and 900 lines of MonotypeDict<Key>.

Let's talk about MonotypeDict<Key>. First off, there are only three valid values for the type parameter Key: int64_t, StringData*, and LowStringPtr. We have a battery of helpers, like Key coerceKey<K>(K k), hashKey<Key>(Key key), and keysEqual<Key>(Key key1, Key key2), to let us deal with these keys generically. (If we ever used an incorrect Key value, we'd get static asserts in these helpers.)

coerceKey returns a tombstone Key if K cannot be coerced (e.g. if Key is LowStringPtr and we call GetInt, or GetStr-with-a-str-with-no-static-counterpart). The string tombstones are nullptr; the int tombstone is the min int64. If you put this int into a MonotypeDict<int64_t>, it'll escalate.

The layout of MonotypeDict<Key> is the same for all Key types. The smallest monodict has a capacity of 6 elements in 128 bytes (compare to the smallest MixedArray: 3 elements in 120 -> 128 bytes):
  16-byte ArrayData header:
      8-byte HeapObject header:
          4-byte refcount
          1-byte ArrayKind
          1-byte GC bits
          1-byte aux bits
          1-byte size index
      4-byte m_size field
      2-byte m_used field
      2-byte bespoke::LayoutIndex
          (low byte stores the DataType!)
  6x16-byte MixedArray<Key>::Elm
      8-byte Key
      8-byte Value
  8x2-byte MixedArray<Key>::Index
      (hash table indices into the Elm array)

As we can see, MonotypeDict is very tightly packed. As a consequence of its design, it can only store up to 1 << 16 elements. I think this tradeoff is acceptable - we can simply escalate to vanilla when it hits this bound. We could pack MonotypeDict slightly more tightly in lowptr builds for the LowStringPtr key case, but I don't think that's worth it - the code would get substantially more complex, and we think structs will do better in this case, anyway.

We have some basic logic in MonotypeDict to maintain the layout above - see sizeIndex(), used(), numElms(), numIndices(), and elmAtIndex(). They just do simple arithmetic to handle these cases. The core of this arithmetic is the scaleForSizeIndex() helper (which was tricky enough that I got it wrong at first).

The next piece of the implementation is the find(Key, hash) API. This API is templated on the operation we're doing, and returns a different result for each operation type:
- find<Add> can only be used when adding a NEW key to the array. It returns a non-null pointer to the hash table index to set.
- find<Get> returns a pointer to the Elm for that key, or nullptr if it's not present.
- find<Remove> returns the position of the hash table index to remove, or -1 if it's not present.
- find<Update> returns an {Elm*, Index*}, where the Elm* may be null but the Index* never is.

The logic of find differs in crucial ways between these operations. For example, Add, Get, and Remove skip over tombstones completely, while Update records the first tombstone index so that it can re-use this index if the element is missing. Add additionally skips key equality checks.

In the hash table, the number of non-empty elements (valid and tombstone) is always bounded by m_used. We can prove this invariant inductively. Set for a new element adds a valid key but increments m_used. Remove converts a valid key to a tombstone (no change to non-empty count) and does not change m_used. Because m_used is always <= numElms(), which is always < numIndices(), there will always be an empty index. We use this property to simplify the termination condition.

After building up all this machinery, the array data operations are mostly careful invocations of find, plus refcounting.

Differential Revision: D24107237

fbshipit-source-id: b85df46b04440489a5da737b6185751f2b7e0fbd