Initial version handshake in json
Summary:
The initial handshake between client and monitor is delicate, since they might be different binary versions. We therefore must be careful to only marshal from one binary into a form that the other binary (maybe newer, maybe older) will accept.

I have long been frustrated by this. I've wanted to send additional information, but I've never been able to.

This diff doesn't achieve anything yet. It's solely an investment into the future. It allows for the monitor to be able to receive a json string from the client instead of just the string version. And it allows for the client to be able to receive a json string back from the monitor instead of just Connection_ok.

My plan is: first, land this right now. Then in two month's time, once no one is ever using the old version any longer, then we can safely start sending additional json information as needed, safe in the knowledge that the receiver will handle it without crashing.

I picked json string because that seems the most flexible. I already implemented this for flow ages ago.

Reviewed By: arxanas, dabek

Differential Revision: D22450521

fbshipit-source-id: b5015e1e90d99aec9611844de18e8033b3719443