Avoid dict/darray appends in builtins
Summary: The more we can avoid this operation, the more we can play games with its implementation. In particular, I'm planning on adding O(n) logging to dict/darray append, and then transitioning it over to be a stateless, O(n) operation. That's not viable for hot usages, but all hot usages are in builtins and can be easily fixed.

Differential Revision: D24366108

fbshipit-source-id: 098ab72cc76d6b2cda63c218cad8e8dcb9783022