clean up the pretty printer situation
Summary:
We created Pp_type because 'deriving show' does not work on GADTs. This makes it sometimes tricky to add other `deriving show`.
Here we inline the pretty printer for the GADT in Typing_defs_core (and all types defined recursively), so that other types can derive show, and remove the Pp_type module.

Differential Revision: D24221724

fbshipit-source-id: 4989e8988f9d853b70d8ee05e7bfbfb0f9419d49