basic support for shapes
Summary:
Shapes are a lot like tuples, but with names entries.

~~In the spirit of simplicity for IFC, in subtyping we only check subtypes for a key when it's present in both shapes. If it's only available in one then we ignore it because the typechecker would have rejected it if it's invalid. This way, we don't have to worry about optional fields and open shapes.~~

We avoid some complexity in typechecking shape literals. Since we are already working with a typechecked AST, we don't have to pair the expected types of each field with the field's expression.

This diff doesn't handle reads and writes to shape fields, I'll save that for later

Reviewed By: mpu

Differential Revision: D24340704

fbshipit-source-id: 2c5ff360d17ef51d43f8def16ff1e6059f0e7f9f