Add TCO for Inferflows and IFC
Summary:
We don't intend on allowing InferFlows in hh_server for www, but InferFlows is very helpful for IFC tests. Since we want both hh_server and hh_single_type_check to consume the same declaration data, it's useful to have declaration be able to support InferFlows.

This diff adds InferFlows as an attribute along with two experimental typechecker features: one for IFC in general and one for inferflows. The first we will likely turn on in hh_server eventually, the second we will likely never allow.

It also changes the attribute nast check to check for sanity on these attributes and raise an error if they are used without the tcopt. It used to be called attribute_arity_check, but it already checked for way more than just attribute arity, so I renamed it.

Reviewed By: zilberstein

Differential Revision: D24267521

fbshipit-source-id: 3b92740bbd3da63c5169df1ab6b5cf81274ef9dd